Implements stricter resource concerns for records and corrects some JSON:API behaviour on sparse fieldsets.
This is a giant and messy commit, but actually comprises few changes of some concepts mangled around the codebase.
Some data structures in `database::schema` have been renamed to be conceptually modelled around resources instead of database tables. Also, there is now a conceptual distinction between attributes and identifiers (which are a subset of attributes).
This reflects on `database::record` -- records now have distinctions between primary keys, foreign keys and general attributes. This is essential for correctly implementing sparse fieldsets and can be leveraged to build optimal queries.
The change in the record format propagates to many parts of the codebase, including the SQLite query builder.
There was also an effort to better name lifetime parameters passed around the codebase; `'sch` for schema lifetime, `'reg` for registry lifetime and `'req` for request lifetime.
There are tests for the data loader, but they test more than they should.